Refactor and robustify dataset structure modification scripts
